+++ src/config.h
@@ -80,6 +80,7 @@ typedef struct {
   bool show_hidden_fonts;
   char font_smoothing;
   char font_render;
+  bool dim_as_font;
   bool bold_as_font;
   bool bold_as_colour;
   bool allow_blinking;