+++ wiki/Changelog.md
@@ -8,6 +8,9 @@ Terminal features
   * Support OSC 52 to clear clibboard (xterm).
   * Fixed DSR Locator status response.
 
+Text rendering
+  * Render dim attribute as a Light font variation if available.
+
 Keyboard handling
   * Revise AltGr handling to support flexible right-Alt+left-Ctrl combinations (#1266).
 
@@ -21,6 +24,7 @@ Configuration
   * New option AllowPasteSelection (#1264).
   * Drop AllowSetSelection from interactive Options menu (~#1264).
   * New option TabFont.
+  * New option DimAsFont.
 
 ### 3.7.1 (24 February 2024) ###